fixing some bugs and making output nicer.

+++ src/main/java/net/dirtcraft/discord/discordlink/Commands/Discord/StopServer.java
@@ -18,9 +18,9 @@ public class StopServer implements DiscordCommandExecutor {
     @Override
     public void execute(MessageSource source, String command, List<String> args) {
         try {
-            if (source.isPrivateMessage()) source.sendCommandResponse("Discord-Link Reboot", "Attempting to reboot the server.");
-            GameChats.getDefaultChat().sendMessage("Discord-Link Reboot", "Attempting to reboot the server.");
-            source.getMessage().delete().queue(s->{},e->{});
+            if (!source.isPrivateMessage()) source.getMessage().delete().queue(s->{},e->{});
+            else source.sendCommandResponse("Forced Reboot Scheduled.", "Attempting to reboot the server.");
+            GameChats.getDefaultChat().sendMessage(source, "Forced Reboot Scheduled.", "Attempting to reboot the server.");
             Thread.sleep(555);
         } catch (Throwable ignored){
